test(macos): run framework windowing probe as app
