+++ .github/workflows/macos-lifecycle.yml
@@ -98,7 +98,8 @@ jobs:
         working-directory: packages/bochki_schedule_app
         run: |
           set -o pipefail
-          flutter test integration_test/macos_framework_windowing_probe_test.dart -d macos --enable-windowing -r expanded 2>&1 | tee framework-windowing-probe.log
+          flutter run -d macos --enable-windowing --dart-define=FRAMEWORK_WINDOWING_PROBE=true -t lib/main.dart 2>&1 | tee framework-windowing-probe.log
+          rg -F '[framework-windowing-probe] PASS child first frame reached main' framework-windowing-probe.log
 
       - name: Upload framework windowing probe log
         if: always()
